requirements.txt
python-dotenv
pathlib
langchain
pygpt4all
qdrant-client
unstructured
sentencepiece
llama-cpp-python==0.1.48
urllib3==1.26.6
pdfminer.six==20221105
python-dotenv==1.0.0
unstructured==0.6.6
extract-msg==0.41.1
tabulate==0.9.0
pandoc==2.3
pypandoc-binary==1.11
fastapi
gunicorn
python-multipart
streamlit
docx2txt
sentence_transformers
pdfminer-six
pre-commit
isort
black
pycln
huggingface-hub
uvicorn
